# LibCrypto: Implement little endian CRC using the slicing-by-8 algorithm