@nico 2023-03-30
Any idea how this compares to the sse4.2 instruction? (Could add `-march=native` to the lagom `gunzip` to get away with a preprocessor check for 4.2. See https://www.corsix.org/content/fast-crc32c-4k for some details – I'd only try doing something basic like I did for arm and hope that that's enough, instead of going super all out.)