@nico 2023-03-30
(i.e. something like

```
#if defined(__SSE4_2__)
  … call __builtin_ia32_crc32qi() / __builtin_ia32_crc32di ...
#endif
```
)