@trflynn89 2023-03-30
> We have a FIXME for that under the AARCH-CRC32 version

Nico wrote that FIXME a few days ago.

Unfortunately the GCC built-ins are hard-coded to use the Castagnoli polynomial, whereas we need to use the Ethernet polynomial for Deflate.