@Hendiadyoin1 2023-03-30
>There are two widely used polynomials, IEEE and Castagnoli. These parameterize the CRC-32 algorithm, and are sometimes known as crc32 and crc32c.
> The IEEE polynomial is used by Bzip2, Ethernet (IEEE 802.3), Gzip, MPEG-2, PNG, SATA, Zip and other formats.
> The Castagnoli polynomial is used by Btrfs, Ext4, iSCSI, SCTP and other formats.
> For better or worse, the hardware instruction computes CRC-32-Castagnoli, which means it's not relevant for e.g. Zip. 
<https://news.ycombinator.com/item?id=19405841>
Ah OK, so until we have SCSI or ext4 or btrfs